requirements.txt
# Core
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
pydantic==2.7.4
pydantic-settings==2.3.4

# Database
sqlalchemy==2.0.31
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Redis
redis==5.0.6
hiredis==2.3.2

# OpenAI + RAG
openai==1.35.0
pinecone-client==4.1.0
tiktoken==0.7.0

# Celery + tasks
celery==5.4.0
flower==2.0.1           # Celery monitoring UI
kombu==5.3.7

# Safety / PII
presidio-analyzer==2.2.354
presidio-anonymizer==2.2.354
spacy==3.7.5

# HTTP
httpx==0.27.0
aiohttp==3.9.5

# Utils
structlog==24.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
rich==13.7.1             # CLI output
typer==0.12.3            # CLI framework
